==> 09b - Process Batter Data.ipynb <==
Add / Process Batter Data

    - Now, we need to process this data and augment our game-level data frame
    - For each game, we want the trailing statistics for each player designated in the starting lineup
    - Need to add the statistics for those 18 players to each game
    - From that, we may want to derive "team-level" averages to simplify our feature set
    - NOTE: this will complicate things when we "go into production" and try and use this model for predicting new games

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

import warnings
warnings.simplefilter(action='ignore', category=pd.errors.PerformanceWarning)

pd.set_option('display.max_columns',5000)
pd.set_option('display.max_rows',5000)

df = pd.read_csv('df_bp7.csv', low_memory=False)
df.shape

    - When aggregating pitcher's past performance, we use default values to "smooth out" players with little / no history
    - With batters, this is more complicated, as different positions have different "typical" hitting capabilities
    - (Separate issue: should we be using training data with pitchers hitting?!?!)

def roll_column(df, col, winsize):
    # do the standard Pandas rolling calc
    t_col = df[col].rolling(winsize, closed='left').sum().to_numpy()
    
    # for the early columns, just do a rolling sum from the beginning
    t_col[:winsize] = np.concatenate(([0],df[col].iloc[:(winsize)].cumsum().to_numpy()[:-1]))

    return(t_col)

## Set up position level defaults

dd = {}
dd_p = {'batavg': .100, 'obp': .150, 'slg': .180, 'slgmod': .220, 'obs': .330, 'sobat': .3}
dd_ss_c = {'batavg': .205, 'obp': .260, 'slg': .300, 'slgmod': .320, 'obs': .540, 'sobat': .25}
dd_2b_3b = {'batavg': .240, 'obp': .280, 'slg': .350, 'slgmod': .355, 'obs': .630, 'sobat': .2}
dd_rest = {'batavg': .255, 'obp': .310, 'slg': .380, 'slgmod': .430, 'obs': .690, 'sobat': .2}
dd['p'] = dd_p
dd['ss'] = dd_ss_c
dd['c'] = dd_ss_c
dd['2b'] = dd_2b_3b
dd['3b'] = dd_2b_3b
dd['1b'] = dd_rest
dd['lf'] = dd_rest
dd['rf'] = dd_rest
dd['cf'] = dd_rest
dd['ph'] = dd_rest
dd['pr'] = dd_ss_c
dd['dh'] = dd_rest

def process_batter_df(p_id, dict_def):
    fname = '/Volumes/CharmedXi/beatVegas/batter_2000/batting_data_'+p_id+'.csv'
    try:
        batter_df = pd.read_csv(fname)
        pos = batter_df.Pos.mode()[0]
        
        # corner cases where the most common position was a pair
        if ',' in pos:
            pos = pos.split(',')[0]
            
        batter_df['date'] = (pd.to_datetime(batter_df.date).astype(str).str.replace('-','')).astype(int)
        t_col = batter_df['dblhead_num'].copy()
        t_col[np.isnan(t_col)] = 0
        batter_df['dblheader_int'] = t_col.astype(int)
        for winsize in [10,30,81,162]:
            suff = str(winsize)
            for raw_col in ['AB','BB','H','2B','3B','HR','HBP','SO','SB','CS']:
                new_col = 'rollsum_'+raw_col+'_'+suff
                batter_df[new_col] = roll_column(batter_df, raw_col, winsize)

            ab_per_game_def = 2
            pa_per_game_def = 2
            batavg_def = dict_def[pos]['batavg']
            obp_def = dict_def[pos]['obp']
            slg_def = dict_def[pos]['slg']
            slgmod_def = dict_def[pos]['slgmod']
            so_bat_perc_def = dict_def[pos]['sobat']

            # Columns created by aggregation above
            ab_col = 'rollsum_AB_'+str(winsize)
            h_col = 'rollsum_H_'+str(winsize)
            bb_col = 'rollsum_BB_'+str(winsize)
            hbp_col = 'rollsum_HBP_'+str(winsize)
            doub_col = 'rollsum_2B_'+str(winsize)
            trip_col = 'rollsum_3B_'+str(winsize)
            hr_col = 'rollsum_HR_'+str(winsize)
            so_col = 'rollsum_SO_'+str(winsize)

            # Columns I will define below
            abmod_col = 'ABmod_'+str(winsize)
            fakeab_col = 'fakeAB_'+str(winsize)
            pa_col = 'PA_'+str(winsize)
            pamod_col = 'PAmod_'+str(winsize)
            fakepa_col = 'fakePA_'+str(winsize)
            xb_col = 'XB_'+str(winsize) # represents extra bases beyond hits
            slg_col = 'SLG_'+str(winsize)
            slgmod_col = 'SLGmod_'+str(winsize)
            batavg_col = 'BATAVG_'+str(winsize)
            so_bat_perc_col = 'SObat_perc_'+str(winsize)
            obp_col = 'OBP_'+str(winsize)
            obs_col = 'OBS_'+str(winsize)

            # calculate BATAVG, with smoothing for low AB numbers
            batter_df[abmod_col] = np.maximum(batter_df[ab_col],winsize*ab_per_game_def)
            batter_df[fakeab_col] = np.minimum(batter_df[abmod_col]-batter_df[ab_col],0)
            batter_df[batavg_col] = (batter_df[h_col] + (batter_df[fakeab_col]*batavg_def))/(batter_df[abmod_col])

            # calculate SLG, with smoothing for low AB numbers
            batter_df[xb_col] = batter_df[doub_col] + 2*batter_df[trip_col] + 3*batter_df[hr_col]
            batter_df[slg_col] = (batter_df[h_col] + batter_df[xb_col] + 
                                     (batter_df[fakeab_col]*slg_def))/(batter_df[abmod_col])

            # calculate OBP, with smoothing for low PA numbers
            batter_df[pa_col] = batter_df[ab_col]+batter_df[bb_col]+batter_df[hbp_col]
            batter_df[pamod_col] = np.maximum(batter_df[pa_col],winsize*pa_per_game_def)
            batter_df[fakepa_col] = np.minimum(batter_df[pamod_col]-batter_df[pa_col],0)
            batter_df[obp_col] = (batter_df[h_col] + batter_df[bb_col] + batter_df[hbp_col]
                                  + (batter_df[fakepa_col]*obp_def))/(
                                    batter_df[pamod_col])

            # calculate SLGmod, with smoothing for low PA numbers
            batter_df[slgmod_col] = (batter_df[so_col] + batter_df[bb_col] + batter_df[hbp_col] 
                                     +batter_df[xb_col] + (batter_df[fakepa_col]*slgmod_def))/(
                                    batter_df[pamod_col])

            # calculate SObat_perc, with smoothing for low PA numbers
            batter_df[so_bat_perc_col] = (batter_df[so_col] + (batter_df[fakepa_col]*so_bat_perc_def))/(
                                    batter_df[pamod_col])

            # calculate OBS
            batter_df[obs_col] = batter_df[obp_col]+batter_df[slg_col]


            batter_df['date_dblhead'] = (batter_df['date'].astype(str) + batter_df['dblheader_int'].astype(str)).astype(int)
            batter_df.set_index('date_dblhead', inplace=True)
    except:
        try:
            print(f'issue for {fname} at position {pos}, returning None')
        except:
            print(f'issue for {fname}, returning None')
        batter_df = None
    return(batter_df)

batter_ids = np.array([])
for num in range(1,10):
    for suffix in ['_h','_v']:
        # Check whether this should be '_id' or '_name'
        colname = 'batter'+str(num)+'_name'+suffix
        batter_ids = np.concatenate((batter_ids, pd.unique(df[colname])))
batter_ids = np.unique(batter_ids)

## Create a batter data dictionary from the saved batter files

batter_data_dict = {}
for i,b_id in enumerate(batter_ids):
    if i % 1000 == 0:
        print(i)
    batter_data_dict[b_id] = process_batter_df(b_id, dd)

new_col_dict = {}
colstems = ['BATAVG', 'OBP', 'SLG', 'OBS', 'SLGmod','SObat_perc']
winsizes = [10,30,81,162]
new_col_list = [stem+'_'+str(winsize)+'_b'+str(i)+hv   for stem in colstems for winsize in winsizes 
                for i in range(1,10) for hv in ['_h','_v']]
for col in new_col_list:
    new_col_dict[col] = np.empty(df.shape[0])
    new_col_dict[col].fill(np.nan)

def get_batter_ids_from_row(row):
    b_cols = ['batter1_name_h', 'batter1_name_v', 'batter2_name_h',
       'batter2_name_v', 'batter3_name_h', 'batter3_name_v',
       'batter4_name_h', 'batter4_name_v', 'batter5_name_h',
       'batter5_name_v', 'batter6_name_h', 'batter6_name_v',
       'batter7_name_h', 'batter7_name_v', 'batter8_name_h',
       'batter8_name_v', 'batter9_name_h', 'batter9_name_v']
    return(row.loc[b_cols].to_dict())

## Add in all the statistics about the starting lineup batters
for i in range(df.shape[0]):
    row = df.iloc[i,:]
    if i % 1000 == 0:
        print(i)
    bid_dict = get_batter_ids_from_row(row)
    date_dblhead = row['date_dblhead']
    for hv in ['_h','_v']:
        for j in range(1,10):
            # Again, check if should be name or id col
            #curr_col = 'batter'+str(j)+'_id'+hv
            curr_col = 'batter'+str(j)+'_name'+hv
            curr_b_id = bid_dict[curr_col]
            if curr_b_id in batter_data_dict.keys():
                curr_batter_df = batter_data_dict[curr_b_id]
                if (curr_batter_df is not None) and (curr_batter_df.shape[0]>0):
                    try:
                        curr_batter_row = curr_batter_df.loc[date_dblhead,:]
                    except:
                        print(f'date not found for batter {curr_b_id} game {date_dblhead}')
                        prev_game_indices = np.where(curr_batter_df.index<date_dblhead)[0]
                        if len(prev_game_indices)==0:
                            index_to_use = 0
                        else:
                            index_to_use = np.max(prev_game_indices)
                        curr_batter_row = curr_batter_df.iloc[index_to_use,:]
                        print(f'using date {curr_batter_df.index[index_to_use]}')
                    if (curr_batter_row.ndim>1):
                        curr_batter_row = curr_batter_row.iloc[0,:]
                    for stem in colstems:
                        for winsize in winsizes:
                            newcolname = stem+'_'+str(winsize)+'_b'+str(j)+hv
                            new_col_dict[newcolname][i] = curr_batter_row[stem+'_'+str(winsize)]
                else:
                    print(f'No data found for {curr_b_id}')
            else:
                print(f'batter not found for {curr_b_id}')

for key, val in new_col_dict.items():
    df[key] = val

df = df.copy()

newcols = new_col_dict.keys()

Deal with missing data

df.loc[:,newcols].isnull().sum()

newcols89 = [stem+'_'+str(winsize)+'_b'+str(i)+hv   for stem in colstems for winsize in winsizes 
                 for hv in ['_h','_v'] for i in range(8,10)]
for col in newcols89:
    stem=col.split('_')[0].lower()
    df[col].fillna(dd['p'][stem])

Create some summary features

w9 = np.array([0.12541131, 0.12159052, 0.11787189, 0.11434144, 0.11096691,
       0.10772781, 0.10430724, 0.10078822, 0.09699465])
w8 = w9[:-1]/np.sum(w9[:-1])

for col in ['BATAVG', 'OBP', 'SLG', 'OBS', 'SLGmod','SObat_perc']:
    for winsize in winsizes:
        for hv in ['_h','_v']:
            b_cols9 = [col+'_'+str(winsize)+'_b'+str(i)+hv for i in range(1,10)]
            b_cols8 = [col+'_'+str(winsize)+'_b'+str(i)+hv for i in range(1,9)]
            fcolname9 = 'lineup9_'+col+'_'+str(winsize)+hv
            fcolname8 = 'lineup8_'+col+'_'+str(winsize)+hv
            fcolname9w = 'lineup9_'+col+'_'+str(winsize)+'_w'+hv
            fcolname8w = 'lineup8_'+col+'_'+str(winsize)+'_w'+hv
            df[fcolname9] = np.mean(df.loc[:,b_cols9].to_numpy(),axis=1)
            df[fcolname8] = np.mean(df.loc[:,b_cols8].to_numpy(),axis=1)
            df[fcolname9w] = df.loc[:,b_cols9].to_numpy().dot(w9)
            df[fcolname8w] = df.loc[:,b_cols8].to_numpy().dot(w8)

df.to_csv('df_bp9.csv', index=False)
